Updating clean architecture - only cover useCase TODO repository and add testings
